Run _.uniqueId calls through parseInt()
Commit http://git.io/9KdxCA changed _.uniqueId so that it will always
return a string.  Previously it would return String if a prefix was
passed, otherwise it would return a number.

The viewer used _.uniqueId for generating ID's for models, which need to be a number
since triple equals is pervasively used for equality checks.

When the main DocumentCloud app upgraded underscore, this broke stuff.